+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## Changed
 - Allow to copy item details of read-only items (do not disable them anymore) - the save and delete button is hidden instead of disabled to avoid user confusion
 - The filter view is focused automatically on overview
+- Automatically focus first element if the filter is active
 
 ### Fixed
 - The case of items is ignored for sorting on item and item authorization screens